+++ lib/poodr_rspec.rb
@@ -1,4 +1,6 @@
 require 'poodr_rspec/version'
+require 'poodr_rspec/gear'
+require 'poodr_rspec/wheel'
 
 module PoodrRspec
 end