+++ lib/poodr_rspec/wheel.rb
@@ -1,0 +1,13 @@
+module PoodrRspec
+  class Wheel
+    attr_reader :rim, :tire
+    def initialize(rim, tire)
+      @rim = rim
+      @tire = tire
+    end
+
+    def diameter
+      rim + (tire * 2)
+    end
+  end
+end